wip on removing pickle files
