+++ app_utils/sections/dataset.py
@@ -25,24 +25,21 @@ from app_utils.utils import (
     get_problem_types,
     get_unique_dataset_name,
     kaggle_download,
-    load_dill,
     local_download,
-    make_label,
     parse_ui_elements,
     remove_temp_files,
     s3_download,
     s3_file_options,
-    save_dill,
 )
 from app_utils.wave_utils import ui_table_from_df
-from llm_studio.src.utils.config_utils import load_config
+from llm_studio.src.utils.config_utils import load_config, make_label
 from llm_studio.src.utils.data_utils import (
     get_fill_columns,
     read_dataframe,
     read_dataframe_drop_missing_labels,
     sanity_check,
 )
-
+from llm_studio.src.utils.utils import load_config_yaml, save_config_yaml
 from .common import clean_dashboard
 
 logger = logging.getLogger(__name__)
@@ -586,7 +583,7 @@ async def dataset_import(
                 # change the default validation strategy if validation df set
                 if cfg.dataset.validation_dataframe != "None":
                     cfg.dataset.validation_strategy = "custom"
-                save_dill(f"{new_path}/{q.client['dataset/import/cfg_file']}.p", cfg)
+                save_config_yaml(f"{new_path}/{q.client['dataset/import/cfg_file']}.yaml", cfg)
 
                 train_rows = None
                 if os.path.exists(cfg.dataset.train_dataframe):
@@ -885,7 +882,7 @@ async def dataset_edit(
     q.client["dataset/import/path"] = dataset.path
     q.client["dataset/import/name"] = dataset.name
     q.client["dataset/import/original_name"] = dataset.name
-    q.client["dataset/import/cfg"] = load_dill(dataset.config_file)
+    q.client["dataset/import/cfg"] = load_config_yaml(dataset.config_file)
 
     if allow_merge and experiments_df.shape[0]:
         allow_merge = False
@@ -961,8 +958,7 @@ async def dataset_display(q: Q) -> None:
         q.client["dataset/display/id"]
     ]
     dataset = q.client.app_db.get_dataset(dataset_id)
-    config_file = dataset.config_file
-    cfg = load_dill(config_file)
+    cfg = load_config_yaml(dataset.config_file)
 
     has_train_df = cfg.dataset.train_dataframe != "None"
