+++ app_utils/utils.py
@@ -25,15 +25,16 @@ import psutil
 from boto3.session import Session
 from botocore.handlers import disable_signing
 from h2o_wave import Q, ui
-from h2o_wave.types import StatListItem
 from pandas.core.frame import DataFrame
 from sqlitedict import SqliteDict
 
 from app_utils.db import Experiment
 from llm_studio.src import possible_values
+from llm_studio.src.utils.config_utils import _get_type_annotation_error, make_label
 from llm_studio.src.utils.data_utils import is_valid_data_frame, read_dataframe
 from llm_studio.src.utils.export_utils import get_size_str
-from llm_studio.src.utils.utils import KNOWN_TYPE_ANNOTATIONS, copy_config
+from llm_studio.src.utils.utils import copy_config, save_config_yaml, load_config_yaml
+from llm_studio.src.utils.type_annotations import KNOWN_TYPE_ANNOTATIONS
 
 from .config import default_cfg
 
@@ -76,13 +77,6 @@ def get_settings_path(q):
     return f"{default_cfg.dbs_path}/{get_user_id(q)}.settings"
 
 
-def _get_type_annotation_error(v: Any, type_annotation: Type) -> ValueError:
-    return ValueError(
-        f"Cannot show {v}: not a dataclass"
-        f" and {type_annotation} is not a known type annotation."
-    )
-
-
 def find_free_port():
     with closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as s:
         s.bind(("", 0))
@@ -103,7 +97,7 @@ def start_process(cfg: Any, gpu_list: List, process_queue: List) -> subprocess.P
     """
 
     num_gpus = len(gpu_list)
-    config_name = f"{cfg.output_directory}/cfg.p"
+    config_name = f"{cfg.output_directory}/cfg.yaml"
 
     if num_gpus == 0:
         p = subprocess.Popen(
@@ -483,22 +477,6 @@ def clean_error(error: str):
     return error
 
 
-def make_label(title: str, appendix: str = "") -> str:
-    """Cleans a label
-
-    Args:
-        title: title to clean
-        appendix: optional appendix
-
-    Returns:
-        Cleaned label
-
-    """
-    label = " ".join(w.capitalize() for w in title.split("_")) + appendix
-    label = label.replace("Llm", "LLM")
-    return label
-
-
 def remove_model_type(problem_type: str) -> str:
     """Removes model type from problem type
 
@@ -622,7 +600,7 @@ def get_dataset(
 
     dataset = dataset.__dict__
 
-    dataset_cfg = load_dill(dataset["config_file"]).dataset.__dict__
+    dataset_cfg = load_config_yaml(dataset["config_file"]).dataset.__dict__
 
     for kk, vv in dataset_cfg.items():
         dataset[kk] = vv
@@ -1178,111 +1156,6 @@ def parse_ui_elements(
     return cfg
 
 
-def get_parent_element(cfg: Any, beautify: bool = True):
-    if hasattr(cfg, "_parent_experiment"):
-        key = "Parent Experiment"
-        value = cfg._parent_experiment
-        if beautify:
-            return ui.stat_list_item(label=key, value=value)
-        return {key: value}
-
-    return None
-
-
-def get_cfg_elements(cfg: Any, q: Q, beautify: bool = True) -> List[StatListItem]:
-    """Returns all single config settings for a given configuration
-
-    Args:
-        cfg: configuration
-        q: Q
-        beautify: flag if the output shall be ran through make_label()
-            strips underscores and Title uppercases.
-
-    Returns:
-        List of stat list items with configuration settings
-    """
-
-    items = []
-
-    parent_element = get_parent_element(cfg, beautify)
-    if parent_element:
-        items.append(parent_element)
-
-    cfg_dict = cfg.__dict__
-    type_annotations = cfg.get_annotations()
-    cfg_dict = {key: cfg_dict[key] for key in cfg._get_order()}
-
-    for k, v in cfg_dict.items():
-
-        if k.startswith("_") or cfg._get_visibility(k) < 0:
-            continue
-
-        if any([x in k for x in ["api"]]):
-            continue
-
-        type_annotation = type_annotations[k]
-
-        if type_annotation in KNOWN_TYPE_ANNOTATIONS:
-            if type_annotation == float:
-                v = float(v)
-            if beautify:
-                t = [ui.stat_list_item(label=make_label(k), value=str(v))]
-            else:
-                t = [{k: v}]
-        elif dataclasses.is_dataclass(v):
-            elements_group = get_cfg_elements(cfg=v, q=q)
-            t = elements_group
-        else:
-            raise _get_type_annotation_error(v, type_annotations[k])
-
-        items += t
-
-    return items
-
-
-def get_grouped_cfg_elements(cfg: Any, q: Q) -> dict:
-    """Returns a grouped config settings dict for a given configuration
-
-    Args:
-        cfg: configuration
-        q: Q
-
-    Returns:
-        Dict of configuration settings
-    """
-
-    cfg_dict = cfg.__dict__
-    type_annotations = cfg.get_annotations()
-    cfg_dict = {key: cfg_dict[key] for key in cfg._get_order()}
-
-    grouped_cfg_dict = {}
-
-    for k, v in cfg_dict.items():
-
-        if k.startswith("_") or cfg._get_visibility(k) < 0:
-            continue
-
-        if any([x in k for x in ["api"]]):
-            continue
-
-        type_annotation = type_annotations[k]
-
-        if type_annotation in KNOWN_TYPE_ANNOTATIONS:
-            grouped_cfg_dict.update({k: v})
-        elif dataclasses.is_dataclass(v):
-            group_items = get_cfg_elements(cfg=v, q=q, beautify=False)
-            group_items = {
-                k: list(v) if isinstance(v, tuple) else v
-                for d in group_items
-                for k, v in d.items()
-            }
-            grouped_cfg_dict.update({k: group_items})
-        else:
-            raise _get_type_annotation_error(v, type_annotations[k])
-
-    return grouped_cfg_dict
-
-
 def get_experiment_status(path: str) -> Tuple[str, str]:
     """Get status information from experiment.
 
@@ -1393,85 +1266,6 @@ def get_experiments_status(df: DataFrame) -> Tuple[List[str], List[str]]:
     return status_all, info_all
 
 
-def load_binary(path: str, backend: Any) -> Any:
-    """Loads a binary file
-
-    Args:
-        path: path of file to load
-        backend: loading backend
-
-    Returns:
-        Loaded object
-    """
-
-    with open(path, "rb") as binary_file:
-        f = backend.load(binary_file)
-        return f
-
-
-def load_dill(path: str) -> Any:
-    """Loads a dill file
-
-    Args:
-        path: path of file to load
-
-    Returns:
-        Loaded dill object
-    """
-
-    return load_binary(path, dill)
-
-
-def load_pickle(path: str) -> Any:
-    """Loads a pickle file
-
-    Args:
-        path: path of file to load
-
-    Returns:
-        Loaded object
-    """
-
-    return load_binary(path, pickle)
-
-
-def save_binary(path: str, obj: Any, backend: Any, kwargs: Dict = {}) -> None:
-    """Saves object as binary file
-
-    Args:
-        path: path of file to save
-        obj: object to save
-        backend: saving backend
-        kwargs: backend kwargs
-    """
-
-    with open(path, "wb") as pickle_file:
-        backend.dump(obj, pickle_file, **kwargs)
-
-
-def save_dill(path: str, obj: Any) -> None:
-    """Saves object as dill file
-
-    Args:
-        path: path of file to save
-        obj: object to save
-    """
-
-    save_binary(path, obj, dill)
-
-
-def save_pickle(path: str, obj: Any, protocol: int = 4) -> None:
-    """Saves object as pickle file
-
-    Args:
-        path: path of file to save
-        obj: object to save
-        protocol: protocol to use when saving pickle
-    """
-
-    save_binary(path, obj, pickle, {"protocol": protocol})
-
-
 def get_experiments_info(df: DataFrame, q: Q) -> DefaultDict:
     """For each experiment in given dataframe, return certain configuration settings
 
@@ -1486,13 +1280,11 @@ def get_experiments_info(df: DataFrame, q: Q) -> DefaultDict:
     info = defaultdict(list)
     for _, row in df.iterrows():
 
+
         try:
-            cfg = load_dill(f"{row.path}/cfg_last.p").__dict__
+            cfg = load_config_yaml(f"{row.path}/cfg.yaml").__dict__
         except Exception:
-            try:
-                cfg = load_dill(f"{row.path}/cfg.p").__dict__
-            except Exception:
-                cfg = None
+            cfg = None
 
         metric = ""
         loss = ""
@@ -1633,7 +1425,7 @@ def get_datasets_info(df: DataFrame, q: Q) -> DefaultDict:
         path = row.path + "/"
 
         try:
-            cfg = load_dill(config_file)
+            cfg = load_config_yaml(config_file)
         except Exception:
             cfg = None
 
@@ -1773,7 +1565,7 @@ def start_experiment(cfg: Any, q: Q, pre: str, gpu_list: Optional[List] = None) 
 
     process_queue = list(set(all_process_queue))
 
-    save_dill(f"{cfg.output_directory}/cfg.p", cfg)
+    save_config_yaml(cfg, f"{cfg.output_directory}/cfg.yaml")
 
     # Start the training process
     p = start_process(cfg=cfg, gpu_list=gpu_list, process_queue=process_queue)
