+++ llm_studio/src/utils/utils.py
@@ -3,15 +3,15 @@ import os
 import pickle
 import random
 import zipfile
-from typing import Any, Tuple
+from typing import Any
 
 import dill
 import numpy as np
 import psutil
 import torch
+import yaml
 
-# types which can be shown directly in the UI without any extra nesting
-KNOWN_TYPE_ANNOTATIONS = [int, float, bool, str, Tuple[str, ...]]
+from llm_studio.src.utils.config_utils import convert_cfg_to_nested_dictionary
 
 logger = logging.getLogger(__name__)
 
@@ -110,21 +110,6 @@ def add_file_to_zip(zf: zipfile.ZipFile, path: str) -> None:
         pass
 
 
-def load_dill(path: str) -> Any:
-    """Loads a dill file
-
-    Args:
-        path: path of file to load
-
-    Returns:
-        Loaded object
-    """
-
-    with open(path, "rb") as binary_file:
-        f = dill.load(binary_file)
-        return f
-
-
 def save_pickle(path: str, obj: Any, protocol: int = 4) -> None:
     """Saves object as pickle file
 
@@ -136,18 +121,6 @@ def save_pickle(path: str, obj: Any, protocol: int = 4) -> None:
 
     with open(path, "wb") as pickle_file:
         pickle.dump(obj, pickle_file, protocol=protocol)
-
-
-def save_dill(path: str, obj: Any) -> None:
-    """Saves object as dill file
-
-    Args:
-        path: path of file to save
-        obj: object to save
-    """
-
-    with open(path, "wb") as pickle_file:
-        dill.dump(obj, pickle_file)
 
 
 def copy_config(cfg: Any) -> Any:
@@ -170,21 +143,27 @@ def copy_config(cfg: Any) -> Any:
     return cfg
 
 
-def save_config(path: str, cfg: Any) -> None:
+def save_config_yaml(path: str, cfg: Any) -> None:
     """Saves config as dill file
 
     Args:
         path: path of file to save to
         cfg: config to save
     """
+    with open(path, "w") as fp:
+        yaml.dump(convert_cfg_to_nested_dictionary(cfg), fp, indent=4)
 
-    # cfg.environment._cpu_comm can't be saved
-    cpu_comm = None
-    if cfg.environment._cpu_comm is not None:
-        cpu_comm = cfg.environment._cpu_comm
-        cfg.environment._cpu_comm = None
-    save_dill(path, cfg)
-    cfg.environment._cpu_comm = cpu_comm
+def load_config_yaml(path: str) -> Any:
+    """Loads config from yaml file
+
+    Args:
+        path: path of file to load from
+    Returns:
+        config object
+    """
+    with open(path, "r") as fp:
+        cfg = yaml.load(fp, Loader=yaml.FullLoader)
+    return cfg
 
 
 class DisableLogger:
